Complete gloo-worker example for spawning WebWorkers
I was not able to figure this out just by reading the documentation,
so I figured this may be worth sharing.

The example is complete in the sense that it provides everything to
run the test application in a browser.

The "controller" spawns a webworker task "worker" and exchanges some
data. The point was to figure out how to start the worker, how to
make it receive the data and how to organize the code.

See README.md for more details.